Move subscribe() method to a separate class an make it non-static [MAILPOET-1825]
